+++ src/kili/services/export/format/voc/__init__.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from typing import Dict, List, Sequence
 from xml.dom import minidom
 
-from kili.orm import JobMLTask, JobTool
+from kili.domain.ontology import JobMLTask, JobTool
 from kili.services.export.exceptions import (
     NoCompatibleJobError,
     NotCompatibleInputType,
